requirements.txt
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tqdm==4.48.2
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.2.0
pandas==1.1.1
python-dotenv==0.14.0
pytesseract==0.3.6
